#948 - Calculates the MP_THOMPSON_*_DATA.DBL tables only during the first execution of the pre-processing.

+++ scripts/2.pre_processing.bash
@@ -100,38 +100,47 @@ fi
 echo ""
 echo -e  "${GREEN}==>${NC} Building MP_THOMPSON DBL tables ...\n"
 
-rm -f ${EXECS}/MP_THOMPSON_*_DATA.DBL
-rm -f ${DATAIN}/fixed/MP_THOMPSON_*_DATA.DBL
-
-cd ${EXECS}
-${EXECS}/build_tables
-
-mv ${EXECS}/MP_THOMPSON_QRacrQG_DATA.DBL    ${DATAIN}/fixed
-mv ${EXECS}/MP_THOMPSON_QRacrQS_DATA.DBL    ${DATAIN}/fixed
-mv ${EXECS}/MP_THOMPSON_freezeH2O_DATA.DBL  ${DATAIN}/fixed
-mv ${EXECS}/MP_THOMPSON_QIautQS_DATA.DBL    ${DATAIN}/fixed
-
-chmod 755 ${DATAIN}/fixed/MP_THOMPSON_*_DATA.DBL
-chgrp $USER ${DATAIN}/fixed/MP_THOMPSON_*_DATA.DBL
-
-# verify here if tables was created ok
 files_needed=("${DATAIN}/fixed/MP_THOMPSON_QRacrQG_DATA.DBL" "${DATAIN}/fixed/MP_THOMPSON_QRacrQS_DATA.DBL" "${DATAIN}/fixed/MP_THOMPSON_freezeH2O_DATA.DBL" "${DATAIN}/fixed/MP_THOMPSON_QIautQS_DATA.DBL")
-for file in "${files_needed[@]}"
-do
-  if [ -s "${file}" ] ; then
-    echo ""
-    echo -e "${GREEN}==>${NC} File ${file} generated sucessfully in ${EXECS} and moved to ${DATAIN}/fixed !"
-    echo
-  else
-    echo -e  "\n${RED}==>${NC} ***** ATTENTION *****\n"   
-    echo -e  "${RED}==>${NC} [${0}] An error occurred during MP_THOMPSON build_tables. At least the file ${file} was not generated. \n"
-    exit -1
-  fi
-done
+
+if [ ! -s ${DATAIN}/fixed/MP_THOMPSON_QRacrQG_DATA.DBL ]
+then
+   echo -e  "${GREEN}==>${NC} This calculation can take around 2 minutes on a supercomputer...\n"
+
+   rm -f ${EXECS}/MP_THOMPSON_*_DATA.DBL
+   rm -f ${DATAIN}/fixed/MP_THOMPSON_*_DATA.DBL
+
+   cd ${EXECS}
+   ${EXECS}/build_tables
+
+   mv ${EXECS}/MP_THOMPSON_QRacrQG_DATA.DBL    ${DATAIN}/fixed
+   mv ${EXECS}/MP_THOMPSON_QRacrQS_DATA.DBL    ${DATAIN}/fixed
+   mv ${EXECS}/MP_THOMPSON_freezeH2O_DATA.DBL  ${DATAIN}/fixed
+   mv ${EXECS}/MP_THOMPSON_QIautQS_DATA.DBL    ${DATAIN}/fixed
+
+   chmod 755 ${DATAIN}/fixed/MP_THOMPSON_*_DATA.DBL
+   chgrp $USER ${DATAIN}/fixed/MP_THOMPSON_*_DATA.DBL
+
+   # verify here if the tables were created 
+   for file in "${files_needed[@]}"
+   do
+     if [ -s "${file}" ] ; then
+       echo ""
+       echo -e "${GREEN}==>${NC} File ${file} generated sucessfully in ${EXECS} and moved to ${DATAIN}/fixed!"
+       echo
+     else
+       echo -e  "\n${RED}==>${NC} ***** ATTENTION *****\n"   
+       echo -e  "${RED}==>${NC} [${0}] An error occurred during MP_THOMPSON build_tables. At least the file ${file} was not generated. \n"
+       exit -1
+     fi
+   done
+else
+   echo -e "${GREEN}==>${NC} MP_THOMPSON DBL tables already exist in ${DATAIN}/fixed!"
+fi
+
 
 # Copying NoahmpTable.TBL from source to datain folder
 echo ""
-echo -e "${GREEN}==>${NC} Copying NoahmpTable.TBL from source to datain fixed folder ...\n"
+echo -e "${GREEN}==>${NC} Copying NoahmpTable.TBL from source code to datain fixed folder ...\n"
 
 if [ ! -s ${DATAIN}/fixed/NoahmpTable.TBL ]
 then
